Split the money steps and fail loudly on a missing booking
The refund steps become three story-focused files — the refund itself, the
extra charges, and the money shares — so no step module runs long and a
mutation run over one story stays narrow.

Reading a booking after checkout now goes through soleBookingOn, which throws
with the count when a listing does not hold exactly one, instead of quietly
taking the first row.
